+++ app/src/main/java/org/simple/clinic/drugs/selection/PrescribedDrugsScreenKey.kt
@@ -3,17 +3,14 @@ package org.simple.clinic.drugs.selection
 import android.os.Parcelable
 import kotlinx.android.parcel.IgnoredOnParcel
 import kotlinx.android.parcel.Parcelize
-import org.simple.clinic.R
-import org.simple.clinic.router.screen.FullScreenKey
+import org.simple.clinic.navigation.v2.ScreenKey
 import java.util.UUID
 
 @Parcelize
-data class PrescribedDrugsScreenKey(val patientUuid: UUID) : FullScreenKey, Parcelable {
+data class PrescribedDrugsScreenKey(val patientUuid: UUID) : ScreenKey(), Parcelable {
 
   @IgnoredOnParcel
   override val analyticsName = "Patient Drugs"
 
-  override fun layoutRes(): Int {
-    return R.layout.screen_patient_prescribed_drugs_entry
-  }
+  override fun instantiateFragment() = EditMedicinesScreen()
 }